refactor: 도메인 별로 예외를 관리하도록 패키지 수정
