+++ src/main/java/roomescape/auth/exception/NoAdminPrivilegeException.java
@@ -1,8 +1,9 @@
-package roomescape.exception;
+package roomescape.auth.exception;
 
 import org.springframework.http.HttpStatus;
 
 public class NoAdminPrivilegeException extends RuntimeException {
+
     public static final int STATUS_CODE = HttpStatus.FORBIDDEN.value();
 
     public NoAdminPrivilegeException(final String message) {
