+++ src/main/java/roomescape/config/AdminCheckInterceptor.java
@@ -6,7 +6,7 @@ import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
 import roomescape.auth.jwt.JwtTokenProvider;
-import roomescape.exception.NoAdminPrivilegeException;
+import roomescape.auth.exception.NoAdminPrivilegeException;
 import roomescape.auth.dto.LoginMember;
 import roomescape.util.CookieUtils;
