--- src/main/java/roomescape/exception/PaymentFailException.java
@@ -1,14 +1,0 @@
-package roomescape.exception;
-
-public class PaymentFailException extends RuntimeException {
-    private final String code;
-
-    public PaymentFailException(String code, String message) {
-        super(message);
-        this.code = code;
-    }
-
-    public String getCode() {
-        return code;
-    }
-}
