+++ src/test/java/roomescape/auth/jwt/JwtTokenProviderTest.java
@@ -11,8 +11,8 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import roomescape.auth.domain.Role;
-import roomescape.exception.ExpiredTokenException;
-import roomescape.exception.UnauthenticatedUserException;
+import roomescape.auth.exception.ExpiredTokenException;
+import roomescape.auth.exception.UnauthenticatedUserException;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberName;
 import roomescape.auth.dto.LoginMember;
